fixed some circular dependency BS
