+++ embodiments/godot-bridge/FEAGIByteStructures/AbstractByteStructure.py
